requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-decouple>=3.8
PyJWT>=2.8.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.5
openpyxl>=3.1.2
xlrd==1.2.0
pandas>=2.1.4
